Fix path for full game image
Include an if statement that will check whether or not each game link is for a bundle page or a game page.

Target the appropriate image class depending on which page the game link is for.

Fix a bug that matched the wrong images to games because links directing to a bundle page were being ignored.