Add admin variable to several @app.route functions
Add Admin Controls navlink to navbar in base template.

The admin variable checks whether or not a user is an admin by checking the boolean attribute of 'admin' within the db. Adding this to each function that generates a new url route ensures that the Admin Controls navlink will appear on each page.