Update app.py file and HTML templates
Update the name of the collection that the application is looking for within MongDB.

As now only one collection handles all of the game data, the same collection (all_pc_games) will be called for each page.

This reduces the amount back-end code and reduces the number of collections within MongoDB, making for a cleaner, more efficient way of storing data.