Update data_steamstore.py for adding data to the MongDB
Update the way in which the data is handled in MongoDB.

No need for separate collections. All games are now stored in a single collection, which includes the following categories:

* Bestsellers

* Award winners

* Action games

* Adventure games

* RPG games

* Strategy games

* Multiplayer games

* Favourites games

Create a new boolean variable for each document which will define what category (or categories) the game belongs to.

Add checks to ensure that no games with the same game title are added as a new document to the collection in order to prevent duplicates, which will fix issues relating to repeating game data.

Update add to db function so that only *one* function handles all the data and adds that data to the db.

Update the 'update' method from 'update' to 'update_one' to prevent any errors from occuring in the terminal, which related to the update method being outdated.