Add content for when there are no review search results
Add content to inform the user that there are no available reviews matching the game that they are looking for.

This is relevant for either the search bar feature, for searching for reviews via game titles, or by navigating to the Reviews page via the Favourites page by clicking on a 'See Reviews' btn for a certain game.

Update the image path for the 'no results' content and link the Request Game btn with the Request Game form.