Update Request A Game to add requests in lowercase
Add the game request string in lowercase letters to ensure that the same games are not being added as a new document just because a character is a different case.

E.g. 'Oblivion', 'OBLIVION', or 'oblivion' would be added to the db as 'oblivion', so if a user tries to add 'OBLIVION' their request would not be added as 'oblivion' already exists.