Update autocomplete feature for Secondary Nav searchbar
Remove data.js file and Export to JS file function in app.py.

The local data.js file is replaced with a more dynamic approach for handling the data in the autocomplete feature.

Define gameData JS variable for the Games pages' autocomplete in games-templates.html to replace gameData in data.js.

Define another JS variable, reviewData, in reviews-template.html in order to add autocomplete functionality to the searchbar on the Reviews page.

Add a new JS file, autocomplete-reviews.js, for handling autocomplete on the Reviews page in order to prevent console errors.

Remove Jinja block content for the searchbar on each game genre page as the repeated code is not necessary.