Fix 'OperationFailure' error for game search results
Fix 'pymongo.errors.OperationFailure: "$search" had the wrong type', caused by clicking the like button on a game card in the search results.

The error triggered once rendering the games-search_results.html' template when 'query' was equal to None.

Fix the error by redirecting the user to the default games page, when the search query is equal to none (once the like button is clicked).