Update Reviews carousel for game review results
Show only the relevant game image in the carousel for when 'See Reviews' button is clicked, or when a game card on the Homepage is clicked.

E.g. Show the Red Dead Redemption 2 image when 'See Reviews' button is clicked on the `Red Dead Redemption` game card on the Favourites page, Game pages and/or Homepage.

This makes it clear to the user what results are being shown to them.