Fix console error for Edit Review page
Fix console error caused by the gameData variable for the autocomplete feature not being defined.

Autocomplete is not needed for the Edit Review form as user will be changing existing reviews, so will not need to search for games.

Disable the 'Search For Game' searchbar to prevent users from changing the game title of their existing review.