Update Request A Game form functionality
Users can no longer submit any string via the Request A Game form.

Update Request A Game form to check if the game title exists within the steam store website, in order to prevent compatability issues.

Ask users to confirm the request submission via a modal popup, once  a match is found. Notify them if the match is different to their request.

Notify users if there game request cannot be submitted due to it not exisiting within the Steam Store.