Refactor Code for Submit Review function
Organize the code that handles submitting user reviews into one function.

Reduce the amount of code in app.py and greatly minimize the amount of duplicate code.

Add the game title for the review to the app.route URL path if the user is navigating to the Submit Review form from their Profile page.